TILT block parsing for all implemented parsers.
TILT=INCLUDE and TILT=NONE are handled differently.
TILT=<filename> was not implemented and is to be considered as an
unhandled event. If a TILT file is in the IES file, the parsing will
ignore it as if TILT=NONE.

Parsed data from the TILT block is stored in a
std::optional<TILT_Data>.

Note that, once data is parsed, it can be modified by the user since a
reference is returned.