Parsing working.
The parser reads each handled standard format and stores its data in a
IES_Document object. The object can be modified freely after that.

TODO: implement a print() function that can be called in order to print
data in a human-readable format.
TODO: try and refactor document out of parsing functions.
TODO: create a wrapper for the string_view vector representing lines.