Printing parsed data working.
Printing is implemented through a IES_Printer static class.
Calling the static method <print> allows to print data from the IES
document in a human-readable format.
The candela values of the luminaire can be printed out to a .csv file,
so they can be elaborated by another program or read in a more comfortable
way.

TODO: try and refactor document out of parsing functions.
TODO: create a wrapper for the string_view vector representing lines.