Implementation of <Label_Data>
Basic implementation of <Label_Data> class, so that the parser is still
working with a new interface for accessing and adding labels in a IES
document object.
Using <Label_Data> to wrap labels allows to access and manipulate label
data in a IES document in an easier way.
Moreover, the container may be changed or updated by modifying the
wrapper instead of all its usages.